requirements.txt
# ============================================
# DEPENDENCIAS DEL PROYECTO DE TRADING
# ============================================

# --- Core Data Science ---
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# --- Visualización ---
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# --- Machine Learning ---
scikit-learn>=1.3.0
tensorflow>=2.13.0  # Para LSTM
keras>=2.13.0

# --- Series Temporales ---
statsmodels>=0.14.0
prophet>=1.1.4  # Facebook Prophet
arch>=6.2.0  # Para modelos GARCH

# --- MetaTrader 5 ---
MetaTrader5>=5.0.45

# --- Web Scraping (para calendario económico) ---
beautifulsoup4>=4.12.0
lxml>=4.9.0

# --- Configuración ---
pyyaml>=6.0
python-dotenv>=1.0.0

# --- Excel/CSV ---
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# --- Utilidades ---
tqdm>=4.65.0  # Barras de progreso
joblib>=1.3.0  # Paralelización

# --- Testing (opcional) ---
pytest>=7.4.0
pytest-cov>=4.1.0

# --- Type Checking (opcional) ---
mypy>=1.5.0